requirements.txt
Pyrogram==2.0.106
TgCrypto==1.2.5
python-dotenv==1.2.2
rich==14.3.3
textual>=0.85.0
psutil>=5.9.0
